The byte returns, and every n by what it counts
glyph_of hands back a glyph index, game_random AH, bonus_kind a kind and
brick_tag a cell value - four byte registers returned as uint32_t.
game_random's limit is DL and goes with it.

`n` was eighteen locals meaning eight different things. particle_count is
a uint16_t field so its five are; hit_count, teleports and player_count
are uint8_t fields so theirs are; the shift and row counters are small;
and the curtain's is a 16-bit product, which is what the `& 0xffff`
around it was saying.

One stays wide and the compiler is why. The cheat screen's `n` indexes a
256-byte buffer this port owns rather than a register the original has,
and as a uint8_t the `n < sizeof line` guarding it can never be false -
-Wtype-limits said so. A bounds check that cannot fail is one that stops
protecting the moment the buffer shrinks, so it keeps its width and the
comment says which of the two kinds of number it is.

Checked: draw_char, draw_sprite_20x6, probe_cell_at, drop_duplicate_hits,
bonus_points and game_random agree with the original, and 2,000 frames on
panel 8 and 1,200 on level 50 are identical throughout.

+++ reconstruct/src/game.h
@@ -1604,7 +1604,7 @@ void paddle_row_offsets(uint32_t x, paddle_rows_t *rows); /* 1ac2:22de */
 void blit_xor(const uint8_t *pixels, const paddle_rows_t *rows); /* 1ac2:2281 */
 void draw_paddle(const uint8_t *sprite);                /* 1ac2:221a */
 void draw_char(char c, uint16_t di);              /* 1ac2:0c64 */
-uint32_t game_random(uint32_t ticks, uint32_t limit);   /* 1ac2:40c0 */
+uint8_t  game_random(uint32_t ticks, uint8_t limit);    /* 1ac2:40c0 */
 void speaker_on(void);                                  /* 1ac2:0085 */
 void speaker_off(void);                                 /* 1ac2:0090 */
 void sound_tick(void);                                  /* 1ac2:0097 */
